+++ src/views/layouts/footer.php
@@ -55,6 +55,5 @@
     </div>
   </footer>
 
-  <script src="<?= e(asset('/js/app.js')) ?>" defer></script>
 </body>
 </html>